Changes to make IE usable within other applications
Creating a RegisterRoutes function that takes a FHIRServer and other URLs. This
will let other applications include IE if they want to run it as part of a
larger application.
